📦 build(version): 5.1.5.30

See merge request tetherfi_product/products/tmac/ui/agent-desktop!56

+++ src/app/modules/t-widgets/tw-content/twc-interaction/twc-interaction.component.ts
@@ -61,7 +61,7 @@ export class TwcInteractionComponent extends TWContentWrapper implements OnInit,
         private _tmacEventService: TMACEventService,
         private _aotWidgetService: AOTWidgetService
     ) {
-        super(hostElement, contentPageService);
+        super('TwcInteractionComponent', hostElement, contentPageService);
     }
 
     /**
@@ -316,6 +316,7 @@ export class TwcInteractionComponent extends TWContentWrapper implements OnInit,
         this.interactions.forEach((i) => {
             if (i.interactionDetails.InteractionID === evt.InteractionID) {
                 i.widgets.aot.forEach((widget) => {
+                    if (!widget.Config.Enabled) return;
                     this._aotWidgetService.destroyWidget(widget.ID);
                 });
             }
